requirements.txt
# Academic Planner - Django Requirements

# Core Django Framework
Django==4.2.24
djangorestframework==3.14.0

# Database Support
psycopg2-binary==2.9.9  # PostgreSQL support (optional)

# AI Integration
google-generativeai==0.3.2

# Environment Management
python-dotenv==1.0.0

# Development Tools (optional)
django-debug-toolbar==4.2.0  # For development debugging
django-extensions==3.2.3     # Additional Django utilities

# Production Dependencies
gunicorn==21.2.0            # WSGI server for production
whitenoise==6.6.0           # Static file serving
dj-database-url==2.1.0      # Database URL parsing

# Security
django-cors-headers==4.3.1  # CORS support for API

# Date/Time Handling
python-dateutil==2.8.2

# JSON Handling
simplejson==3.19.2

# HTTP Requests
requests==2.31.0

# File Handling
Pillow==10.1.0              # Image processing (if needed)

# Testing (optional)
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality (optional)
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentation (optional)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
